sales CUF accessibility screen I design
code clean-up

Added the form type variable for sale,support and accessibility forms and also fixed the content issue

Updated the sale accessibility page contents.

Added the validation for accessibility form.

Added the validation and removed the unused code

Fixed the UI issue and added the form validation

Added the validation for pdfAccessibility Form

Integrated the API with pdfAccessibility form.

Updated the URL